[NOT FOR MERGE] Add common headers that reverse proxies set
These proxy headers are often set by reverse proxies, or web servers like nginx
that are capable of proxying requests to a backend application server that runs
SimpleSAMLphp.

I've marked this as 'not for merge' because theoretically anyone could set these
headers, and we want to make sure that they're only set by trusted proxies. I
propose using an environment variable to enumerate a list of trusted IP
addresses, and setting a define() with a boolean true/false for a quick check
point, e.g. `if(TRUSTED_PROXY && isset(...`, however I don't know how this
should be handled in SimpleSAMLphp.

Are there any thoughts on how this could be handled better? Ideally we wouldn't
hard-code things in config.php, because these would likely vary per-environment
(development environments may not have them, and trusted proxies are likely
different IP addresses in staging/pre-prod environments and production).

Any thoughts on how to solve this properly?